requirements.txt
requests
telebot
openai
PyYAML
python-telegram-bot==13.7
flask
python-dotenv
